server-py/requirements.txt
opencv-python>=4.6.0
pyzmq>=24.0.1
numpy>=1.23.4
